+++ views/user/stats_templ.go
@@ -201,7 +201,7 @@ func StatsRecordDistanceDate(workout *database.Float64Record) templ.Component {
 	})
 }
 
-func StatsRecordDate(id uint, date *time.Time) templ.Component {
+func StatsRecordDate(id uint64, date *time.Time) templ.Component {
 	return templruntime.GeneratedTemplate(func(templ_7745c5c3_Input templruntime.GeneratedComponentInput) (templ_7745c5c3_Err error) {
 		templ_7745c5c3_W, ctx := templ_7745c5c3_Input.Writer, templ_7745c5c3_Input.Context
 		if templ_7745c5c3_CtxErr := ctx.Err(); templ_7745c5c3_CtxErr != nil {
